config.py
import os
from dotenv import load_dotenv

basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir, ".env"))


class Config(object):
    SECRET_KEY = os.environ.get("SECRET_KEY") or "some-secret"
    SQLALCHEMY_DATABASE_URI = os.environ.get("DATABASE_URL") or \
        "sqlite:///" + os.path.join(basedir, "app.db")
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    # JWT_ACCESS_TOKEN_EXPIRES=os.environ.get("TOKEN_EXPIRATION_SECONDS") or 86400


class DevelopmentConfig(Config):
    pass


class ProductionConfig(Config):
    pass
